shouldbeExtension for newlines
