+++ src/Stryker.Core/Stryker.Core.UnitTest/Mutants/MutantorchestratorTests.cs
@@ -3,6 +3,7 @@ using Microsoft.CodeAnalysis.CSharp;
 using Shouldly;
 using Stryker.Core.Mutants;
 using Stryker.Core.Mutators;
+using Stryker.Core.UnitTest;
 using System;
 using System.Collections.ObjectModel;
 using System.IO;
@@ -50,33 +51,6 @@ namespace StrykerNet.UnitTest.Mutants
             var result = _target.Mutate(tree.GetRoot());
 
             _target.GetLatestMutantBatch().Count().ShouldBe(numberOfMutations);
-        }
-    }
-
-    /// <summary>
-    /// Making asserting syntax trees more easy
-    /// </summary>
-    public static class AssertionExtensions
-    {
-        /// <summary>
-        /// Compares two code strings and asserts equality 
-        /// </summary>
-        /// <param name="actual">Resulted code</param>
-        /// <param name="expected">Expected code</param>
-        public static void ShouldBeSemantically(this string actual, string expected)
-        {
-            ShouldBeSemantically(CSharpSyntaxTree.ParseText(actual).GetRoot(), CSharpSyntaxTree.ParseText(expected).GetRoot());
-        }
-
-        /// <summary>
-        /// Compares two syntax trees and asserts equality 
-        /// </summary>
-        /// <param name="actual">Resulted code</param>
-        /// <param name="expected">Expected code</param>
-        public static void ShouldBeSemantically(this SyntaxNode actual, SyntaxNode expected)
-        {
-            SyntaxFactory.AreEquivalent(actual, expected)
-                .ShouldBeTrue($"AST's are not equavalent. Actual: {Environment.NewLine}{actual}, expected: {Environment.NewLine}{expected}");
         }
     }
 }
